feat: integrate Kafka for Steam import, add achievements translations and initialization, restructure files
- Integrated Kafka for Steam import
- Moved all DTOs into a single package
- Added translations (ru, en) for achievements
- Implemented achievements initialization
- Slightly updated file organization

BREAKING CHANGE: DTO package structure has changed

+++ src/main/java/com/gamevault/service/GameService.java
@@ -8,7 +8,7 @@ import com.gamevault.db.repository.GameRepository;
 import com.gamevault.exception.GameNotFoundInIgdbException;
 import com.gamevault.exception.IgdbFetchException;
 import com.gamevault.exception.IgdbParsingException;
-import com.gamevault.dto.igdb.IgdbGameDTO;
+import com.gamevault.dto.output.igdb.IgdbGameDTO;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -44,7 +44,7 @@ public class GameService {
             IgdbGameDTO igdbGameDto = games.get(0);
 
             Game game = new Game();
-            game.setIgdbId((long) igdbGameDto.id());
+            game.setIgdbId(igdbGameDto.id());
             game.setTitle(igdbGameDto.name());
             game.setDescription(igdbGameDto.summary());
             game.setCoverUrl(igdbGameDto.cover().url());
